+++ lib/ops/operatoracl.go
@@ -571,13 +571,6 @@ func (o *OperatorACL) CreateProgressEntry(key SiteOperationKey, entry ProgressEn
 	return o.operator.CreateProgressEntry(key, entry)
 }
 
-func (o *OperatorACL) GetSiteOperationCrashReport(key SiteOperationKey) (io.ReadCloser, error) {
-	if err := o.ClusterAction(key.SiteDomain, storage.KindCluster, teleservices.VerbRead); err != nil {
-		return nil, trace.Wrap(err)
-	}
-	return o.operator.GetSiteOperationCrashReport(key)
-}
-
 func (o *OperatorACL) GetSiteReport(key SiteKey) (io.ReadCloser, error) {
 	if err := o.ClusterAction(key.SiteDomain, storage.KindCluster, teleservices.VerbRead); err != nil {
 		return nil, trace.Wrap(err)