+++ lib/report/bash.go
@@ -31,7 +31,7 @@ import (
 
 // Collect fetches shell histories for all users from passwd.
 // Collect implements Collector
-func (r bashHistoryCollector) Collect(ctx context.Context, reportWriter Writer, runner utils.CommandRunner) error {
+func (r bashHistoryCollector) Collect(ctx context.Context, reportWriter FileWriter, runner utils.CommandRunner) error {
 	log.Debug("collecting bash histories")
 	passwd, err := utils.GetPasswd()
 	if err != nil {
@@ -45,7 +45,7 @@ func (r bashHistoryCollector) Collect(ctx context.Context, reportWriter Writer, 
 	}
 
 	for _, user := range users {
-		w, err := reportWriter(fmt.Sprintf("bash_history-%v", user.Name))
+		w, err := reportWriter.NewWriter(fmt.Sprintf("bash_history-%v", user.Name))
 		if err != nil {
 			log.Warningf("failed to create writer for bash history for user %q", user.Name)
 			continue