+++ lib/report/k8s.go
@@ -25,13 +25,12 @@ import (
 	"github.com/gravitational/gravity/lib/utils"
 	"github.com/gravitational/gravity/lib/utils/kubectl"
 
-	"github.com/gravitational/trace"
 	log "github.com/sirupsen/logrus"
 )
 
-// KubernetesInfo returns a list of collectors to fetch kubernetes-related
+// NewKubernetesCollector returns a list of collectors to fetch kubernetes-specific
 // diagnostics.
-func KubernetesInfo(ctx context.Context, runner utils.CommandRunner) Collectors {
+func NewKubernetesCollector(ctx context.Context, runner utils.CommandRunner) Collectors {
 	runner = planetContextRunner{runner}
 	// general kubernetes info
 	commands := Collectors{
@@ -49,7 +48,6 @@ func KubernetesInfo(ctx context.Context, runner utils.CommandRunner) Collectors 
 	if err != nil || len(namespaces) == 0 {
 		namespaces = defaults.UsedNamespaces
 	}
-	log.Debugf("kubernetes namespaces: %v", namespaces)
 
 	for _, namespace := range namespaces {
 		for _, resourceType := range defaults.KubernetesReportResourceTypes {
@@ -58,17 +56,20 @@ func KubernetesInfo(ctx context.Context, runner utils.CommandRunner) Collectors 
 				utils.PlanetCommand(kubectl.Command("describe", resourceType, "--namespace", namespace))...))
 		}
 
+		logger := log.WithField("namespace", namespace)
 		// fetch pod logs
 		pods, err := kubectl.GetPods(ctx, namespace, runner)
 		if err != nil {
-			log.Errorf("failed to query pods in namespace %v: %v", namespace, trace.DebugReport(err))
+			logger.WithError(err).Warn("Failed to query pods.")
 			continue
 		}
 		for _, pod := range pods {
 			containers, err := kubectl.GetPodContainers(ctx, namespace, pod, runner)
 			if err != nil {
-				log.Errorf("failed to query container in pod %v in namespace %v: %v",
-					pod, namespace, trace.DebugReport(err))
+				logger.WithFields(log.Fields{
+					log.ErrorKey: err,
+					"pod":        pod,
+				}).Warn("Failed to query container.")
 				continue
 			}
 			for _, container := range containers {