+++ docs/5.x/changelog.md
@@ -8,8 +8,8 @@ LTS starts with `3.51.0` with minor backwards compatible changes added over time
 | Release       | LTS | Release Date         | Supported Until      | Kubernetes Version | Teleport Version |
 | --------------|-----| -------------------- | -------------------- | ------------------ |------------------|
 | 6.0.0-rc.2    | -   | June 13th, 2019      | -                    | 1.14.1             | 3.2.5            |
-| 5.6.1         | No  | April 17th, 2019     | -                    | 1.14.0             | 3.0.5            |
-| 5.5.9         | Yes | June 6th, 2019       | September 7th, 2020  | 1.13.5             | 3.0.4            |
+| 5.6.2         | No  | June 18th, 2019      | -                    | 1.14.0             | 3.0.5            |
+| 5.5.10        | Yes | June 17th, 2019      | September 7th, 2020  | 1.13.5             | 3.0.4            |
 | 5.4.10        | No  | March 26th, 2019     | -                    | 1.13.5             | 2.4.10           |
 | 5.3.9*        | No  | March 7th, 2019      | -                    | 1.12.3             | 2.4.7            |
 | 5.2.12        | Yes | March 26th, 2019     | October 15th, 2019   | 1.11.9             | 2.4.10           |
@@ -70,6 +70,19 @@ LTS starts with `3.51.0` with minor backwards compatible changes added over time
 * Add support for using `helm` directly from host.
 
 ## 5.x Releases
+
+### 5.6.2
+
+#### Improvements
+
+* Add support for using `helm` directly from host.
+
+### 5.5.10 LTS
+
+#### Bugfixes
+
+* Fix an issue with automatic NO_PROXY rules that break some cluster operations.
+* Fix an issue with updating or removing environment configuration.
 
 ### 5.5.9 LTS